requirements.txt
python-kasa
fastapi